Move thread manager to extras and generalize the class
